requirements.txt
absl-py==0.5.0
backcall==0.1.0
bleach==1.5.0
cycler==0.10.0
decorator==4.3.0
h5py==2.8.0
html5lib==0.9999999
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
jedi==0.13.1
jupyter-client==5.2.3
jupyter-core==4.4.0
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
numpy==1.15.2
open3d-python==0.3.0.0
opencv-contrib-python==3.4.2.17
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
pkg-resources==0.0.0
prompt-toolkit==2.0.5
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
simplegeneric==0.8.1
six==1.11.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
tornado==5.1.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
